+++ CHANGELOG.md
@@ -3,7 +3,17 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/alexdlaird/hookee/compare/1.2.0...HEAD)
+## [Unreleased](https://github.com/alexdlaird/hookee/compare/1.2.1...HEAD)
+
+## [1.2.1](https://github.com/alexdlaird/hookee/compare/1.2.0...1.2.1) - 2019-09-17
+### Added
+- Documentation improvements.
+
+### Changed
+- Renamed `Config.click_ctx` to `Config.click_logging`, is now kwarg to `Config` so it can be passed on instantiation.
+
+### Fixed
+- Formatting in console output.
 
 ## [1.2.0](https://github.com/alexdlaird/hookee/compare/1.1.0...1.2.0) - 2019-09-16
 ### Added